--- client/src/App/pages/Login/index.jsx
@@ -1,101 +1,0 @@
-import React from 'react'
-import { Typography, Paper, Avatar, Button, FormControl, Input, InputLabel } from '@material-ui/core'
-import LockOutlinedIcon from '@material-ui/icons/LockOutlined'
-import withStyles from '@material-ui/core/styles/withStyles'
-import 'firebase/auth';
-import { auth, provider } from '../../firebaseConfig.js';
-import axios from 'axios';
-
-const styles = theme => ({
-    main: {
-        width: 'auto',
-        marginLeft: theme.spacing.unit * 3,
-        marginRight: theme.spacing.unit * 3,
-        [theme.breakpoints.up(400 + theme.spacing.unit * 3 * 2)]: {
-            width: 400,
-            marginLeft: 'auto',
-            marginRight: 'auto',
-        },
-    },
-    paper: {
-        marginTop: theme.spacing.unit * 8,
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center',
-        padding: `${theme.spacing.unit * 2}px ${theme.spacing.unit * 3}px ${theme.spacing.unit * 3}px`,
-    },
-    avatar: {
-        margin: theme.spacing.unit,
-        backgroundColor: theme.palette.secondary.main,
-    },
-    form: {
-        marginTop: theme.spacing.unit,
-    },
-    submit: {
-        marginTop: theme.spacing.unit * 3,
-    },
-});
-
-const gmailLogin = () => {
-    //use firebase signinwithpopupmethod. takes google auth provider
-    auth.signInWithPopup(provider).then((result) => {
-        //reset state to current logged in user
-        // this.setState({
-        //     user: result.user
-        // })
-        axios.post('/gmailLogin', {
-            result
-        }).then((res) => {
-            console.log('this works');
-        })
-    })
-}
-
-function Login(props) {
-    const { classes } = props
-
-    return (
-        <main className={classes.main}>
-            <Paper className={classes.paper}>
-                <Avatar className={classes.avatar}>
-                    <LockOutlinedIcon />
-                </Avatar>
-                <Typography component="h1" variant="h5">
-                    Sign in
-                </Typography>
-                <form className={classes.form}>
-                    <FormControl margin="normal" required fullWidth>
-                        <InputLabel htmlFor="email">Email Address</InputLabel>
-                        <Input id="email" name="email" autoComplete="off" autoFocus />
-                    </FormControl>
-                    <FormControl margin="normal" required fullWidth>
-                        <InputLabel htmlFor="password">Password</InputLabel>
-                        <Input name="password" type="password" id="password" autoComplete="off" />
-                    </FormControl>
-                    <Button
-                        type="submit"
-                        fullWidth
-                        variant="contained"
-                        color="primary"
-                        className={classes.submit}>
-                        Sign in
-                    </Button>
-                    <Button
-                        type="submit"
-                        fullWidth
-                        variant="outlined"
-                        color="secondary"
-                        className={classes.submit}>
-                        Register
-                    </Button>
-                    <Button 
-                        type="submit"
-                        onClick={gmailLogin}>G-Mail
-                    </Button>
-                </form>
-            </Paper>
-        </main>
-    )
-}
-//
-export default withStyles(styles)(Login)
